S4_Models/S4_requirements.txt
langchain-core
langchain-openai
langchain-community
requests
python-dotenv
together